requirements.txt
PyQt6
opencv-python
pygame
pyserial
pandas
xlsxwriter
paho-mqtt
pyqtdarktheme
